% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.test_dumb.R
\name{t_test_dumb}
\alias{t_test_dumb}
\title{FUNCTION_TITLE}
\usage{
t_test_dumb(.data, value, group, check_assumptions = TRUE,
  var.equal = FALSE, ...)
}
\arguments{
\item{group}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}

\item{response}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
