% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.test_dumb.R
\name{t_test_dumb}
\alias{t_test_dumb}
\title{t test for dummies}
\usage{
t_test_dumb(.data, value, group, check_assumptions = TRUE,
  var.equal = FALSE, ...)
}
\arguments{
\item{group}{the grouping column}

\item{data}{the data to be included}

\item{response}{the values column}
}
\value{
the same object as the t.test with diagnostics included
}
\description{
t test with diagnostics
}
\details{
To be done
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
